+++ adapters/sonoff/__init__.py
@@ -1,0 +1,13 @@
+from adapters.on_off_switch_adapter import OnOffSwitchAdapter
+from adapters.generic.motion_sensor import MotionSensorAdapter
+from adapters.generic.temp_hum_sensor import TemperatureHumiditySensorAdapter
+from adapters.sonoff.SNZB04 import SNZB04
+
+
+sonoff_adapters = {
+    'BASICZBR3': OnOffSwitchAdapter,                # SONOFF Zigbee smart switch
+    'S31ZB': OnOffSwitchAdapter,                    # SONOFF Zigbee smart plug (US version)
+    'SNZB-02': TemperatureHumiditySensorAdapter,    # SONOFF Temperature and humidity sensor
+    'SNZB-03': MotionSensorAdapter,                 # SONOFF Motion sensor
+    'SNZB-04': SNZB04,                              # SONOFF Contact sensor
+}